# Replace readAllAppend and limitedReaderPlus with more efficient readNAppend