--- a/resp/bench_test.go
+++ b/resp/bench_test.go
@@ -2,7 +2,6 @@ package resp
 
 import (
 	"bufio"
-	"bytes"
 	"fmt"
 	"strconv"
 	"strings"
@@ -46,24 +45,25 @@ var bfloat float64
 func BenchmarkReadFloat(b *testing.B) {
 	tests := []struct {
 		In string
+		N int
 	}{
-		{"1"},
-		{"1.23"},
-		{"-1"},
-		{"-1.23"},
-		{"+1"},
-		{"+1.23"},
+		{"1", 1},
+		{"1.23", 4},
+		{"-1", 2},
+		{"-1.23", 5},
+		{"+1", 2},
+		{"+1.23", 5},
 	}
 
 	for _, test := range tests {
-		input := test.In
+		input, n := test.In, test.N
 
 		b.Run(fmt.Sprint(test.In), func(b *testing.B) {
 			var r strings.Reader
 
 			for i := 0; i < b.N; i++ {
 				r.Reset(input)
-				bfloat, _ = readFloat(&r, 64)
+				bfloat, _ = readFloat(&r, 64, n)
 			}
 		})
 	}
@@ -74,24 +74,25 @@ var bint int64
 func BenchmarkReadInt(b *testing.B) {
 	tests := []struct {
 		In string
+		N int
 	}{
-		{"1"},
-		{"123"},
-		{"-1"},
-		{"-123"},
-		{"+1"},
-		{"+123"},
+		{"1", 1},
+		{"123", 3},
+		{"-1", 2},
+		{"-123", 4},
+		{"+1", 2},
+		{"+123", 4},
 	}
 
 	for _, test := range tests {
-		input := test.In
+		input, n := test.In, test.N
 
 		b.Run(fmt.Sprint(test.In), func(b *testing.B) {
 			var r strings.Reader
 
 			for i := 0; i < b.N; i++ {
 				r.Reset(input)
-				bint, _ = readInt(&r)
+				bint, _ = readInt(&r, n)
 			}
 		})
 	}
@@ -102,41 +103,21 @@ var buint uint64
 func BenchmarkReadUint(b *testing.B) {
 	tests := []struct {
 		In string
+		N int
 	}{
-		{"1"},
-		{"123"},
+		{"1", 1},
+		{"123", 123},
 	}
 
 	for _, test := range tests {
-		input := test.In
+		input, n := test.In, test.N
 
 		b.Run(fmt.Sprint(test.In), func(b *testing.B) {
 			var r strings.Reader
 
 			for i := 0; i < b.N; i++ {
 				r.Reset(input)
-				buint, _ = readUint(&r)
-			}
-		})
-	}
-}
-
-func BenchmarkReadAllAppend(b *testing.B) {
-	respBytes := []byte("$5\r\nhello\r\n")
-
-	for _, bcap := range []int{0, len(respBytes), len(respBytes) + bytes.MinRead} {
-		b.Run("Capacity"+strconv.Itoa(bcap), func(b *testing.B) {
-			var br bytes.Reader
-			buf := make([]byte, 0, bcap)
-
-			b.ResetTimer()
-
-			for i := 0; i < b.N; i++ {
-				br.Reset(respBytes)
-
-				if _, err := readAllAppend(&br, buf); err != nil {
-					b.Fatal(err)
-				}
+				buint, _ = readUint(&r, n)
 			}
 		})
 	}
@@ -148,25 +129,16 @@ func (nothingReader) Read(p []byte) (n int, err error) {
 	return len(p), nil
 }
 
-// BenchmarkReadAllAppendLRP benchmarks for the common case
-// where the io.Reader passed to readAllAppend is an
-// *limitedReaderPlus.
-func BenchmarkReadAllAppendLRP(b *testing.B) {
+func BenchmarkReadNAppend(b *testing.B) {
 	for _, n := range []int{0, 64, 512, 4096} {
 		b.Run("N=" + strconv.Itoa(n), func(b *testing.B) {
-			br := bufio.NewReader(nothingReader{})
+			var r nothingReader
 			buf := *getBytes()
-
-			lrp := &limitedReaderPlus{}
-			lrp.lr.R = br
 
 			b.ResetTimer()
 
 			for i := 0; i < b.N; i++ {
-				lrp.eof = false
-				lrp.lr.N = int64(n)
-
-				if _, err := readAllAppend(lrp, buf); err != nil {
+				if _, err := readNAppend(&r, buf, n); err != nil {
 					b.Fatal(err)
 				}
 			}