--- a/resp/util.go
+++ b/resp/util.go
@@ -1,7 +1,6 @@
 package resp
 
 import (
-	"bufio"
 	"io"
 )
 
@@ -31,35 +30,4 @@ func (lr *lenReader) Read(b []byte) (int, error) {
 
 func (lr *lenReader) Len() int64 {
 	return lr.l
-}
-
-// A special limited reader which will read an extra two bytes after the limit
-// has been reached
-type limitedReaderPlus struct {
-	eof bool
-	lr  io.LimitedReader
-}
-
-func (lrp *limitedReaderPlus) reset(br *bufio.Reader, limit int64) {
-	lrp.eof = false
-	lrp.lr = io.LimitedReader{
-		R: br,
-		N: limit,
-	}
-}
-
-func (lrp *limitedReaderPlus) Read(b []byte) (int, error) {
-	if lrp.eof {
-		return 0, io.EOF
-	}
-
-	i, err := lrp.lr.Read(b)
-	// we need to check lrp.lr.N manually since lrp.lr.Read will not
-	// return io.EOF if len(b) == 0.
-	if lrp.lr.N <= 0 {
-		lrp.eof = true
-		_, err = lrp.lr.R.(*bufio.Reader).Discard(2)
-		return i, err
-	}
-	return i, err
 }