When ever `readAllAppend` is called it receives an `*limitedReaderPlus`, which always ends after an specific amount of bytes that we have specified before. This change makes use of this fact and replaces `readAllAppend` and the `limitedReaderPlus` type with a new function `readNAppend` that reads a given number of bytes. This avoids some allocation and memory movement and removes the need for copying into a `*bytes.Buffer` first and also removes an usage of `sync.Pool`.

Benchmark:

```
name                       old time/op    new time/op    delta
BenchmarkReadNAppend/N=0-8       65.8ns ± 0%    18.5ns ± 0%  -71.90%  (p=0.000 n=10+8)
BenchmarkReadNAppend/N=64-8      80.2ns ± 3%    39.8ns ± 0%  -50.34%  (p=0.000 n=10+9)
BenchmarkReadNAppend/N=512-8      253ns ± 2%     206ns ± 2%  -18.64%  (p=0.000 n=9+10)
BenchmarkReadNAppend/N=4096-8    1.09µs ± 3%    1.04µs ± 5%   -5.22%  (p=0.001 n=10+10)

name                       old alloc/op   new alloc/op   delta
BenchmarkReadNAppend/N=0-8        0.00B          0.00B          ~     (all equal)
BenchmarkReadNAppend/N=64-8       0.00B          0.00B          ~     (all equal)
BenchmarkReadNAppend/N=512-8       512B ± 0%      512B ± 0%     ~     (all equal)
BenchmarkReadNAppend/N=4096-8    4.10kB ± 0%    4.10kB ± 0%   -0.02%  (p=0.000 n=10+10)

name                       old allocs/op  new allocs/op  delta
BenchmarkReadNAppend/N=0-8         0.00           0.00          ~     (all equal)
BenchmarkReadNAppend/N=64-8        0.00           0.00          ~     (all equal)
BenchmarkReadNAppend/N=512-8       1.00 ± 0%      1.00 ± 0%     ~     (all equal)
BenchmarkReadNAppend/N=4096-8      1.00 ± 0%      1.00 ± 0%     ~     (all equal)
```