@nussjustin 2018-10-07
The radix GetSet benchmarks show no changes:

```
name                     old time/op    new time/op    delta
SerialGetSet/radix-8       43.9µs ±10%    44.5µs ± 2%    ~     (p=0.897 n=10+8)
ParallelGetSet/radix-8     13.2µs ± 4%    13.4µs ± 6%    ~     (p=0.247 n=10+10)

name                     old alloc/op   new alloc/op   delta
SerialGetSet/radix-8        67.0B ± 0%     67.0B ± 0%    ~     (all equal)
ParallelGetSet/radix-8       100B ± 0%       99B ± 0%  -1.00%  (p=0.000 n=10+9)

name                     old allocs/op  new allocs/op  delta
SerialGetSet/radix-8         4.00 ± 0%      4.00 ± 0%    ~     (all equal)
ParallelGetSet/radix-8       5.00 ± 0%      5.00 ± 0%    ~     (all equal)
```