@nussjustin 2018-10-07
A local test for GET and MGET with big values shows a great improvement of >36%

```
name         old time/op    new time/op    delta
RedisGet-8      106ms ± 7%      62ms ± 5%  -41.17%  (p=0.000 n=10+10)
RedisMGet-8    87.4ms ± 8%    55.8ms ± 7%  -36.15%  (p=0.000 n=10+10)

name         old alloc/op   new alloc/op   delta
RedisGet-8      513MB ±13%     220MB ± 6%  -57.12%  (p=0.000 n=10+10)
RedisMGet-8     353MB ±20%     161MB ± 8%  -54.41%  (p=0.000 n=10+10)

name         old allocs/op  new allocs/op  delta
RedisGet-8        112 ±16%        75 ± 4%  -33.07%  (p=0.000 n=10+10)
RedisMGet-8      52.7 ±15%      32.4 ± 5%  -38.44%  (p=0.000 n=10+9)
```

The code can be found here: https://gist.github.com/nussjustin/7485812ce6303c8980b88b07f26b3474