@mediocregopher 2018-10-08
This looks pretty awesome at first glance, lemme take a better look at it tomorrow and make sure I understand it properly before merging