@nussjustin 2018-10-09
Yeah, while the idea is quiet simple the implementation is quite tricky and took me some time to get right. There's probably still area to improvement, especially with the io.Reader handling (0 byte reading).

But given the performance improvement for my gist, the implementation looked like a great tradeoff when I tested it.