@nussjustin 2018-10-10
Please take a look. Commit 3 to 5 address your comments. The last 2 commits (ok, the second last) removes `readAllAppend` and `limitedReaderPlus` as you suggested.