@nussjustin 2018-10-10
The newest code gives even more improvement on my gist:

```
name         old time/op    new time/op    delta
RedisGet-8      124ms ± 2%      64ms ± 7%  -48.61%  (p=0.000 n=8+10)
RedisMGet-8     101ms ±14%      59ms ± 3%  -41.37%  (p=0.000 n=10+10)

name         old alloc/op   new alloc/op   delta
RedisGet-8      512MB ± 9%     210MB ± 7%  -58.97%  (p=0.000 n=10+10)
RedisMGet-8     350MB ±24%     163MB ± 8%  -53.47%  (p=0.000 n=10+9)

name         old allocs/op  new allocs/op  delta
RedisGet-8        110 ±11%        70 ± 4%  -36.13%  (p=0.000 n=10+9)
RedisMGet-8      52.2 ±23%      31.7 ± 4%  -39.34%  (p=0.000 n=10+9)
```