@mediocregopher 2018-10-11
Nice! Thanks so much for all that, these speed improvements are amazing!