@mediocregopher 2018-10-11
Ack! actually I did have one question, forgot to ask, on line 804:

```
if err := a.unmarshalSingle(br, int(l)); err != nil && err != io.EOF {
```

What's the purpose of allowing `io.EOF`? if that was hit before the the trailing `\r\n` were discarded I would think that error should be passed up